test: share fixtures with conftest.py
Utilize 'conftest.py' for sharing test fixtures, currently isolated
within test modules.
