+++ tests/test_register.py
@@ -1,5 +1,5 @@
 """Test register.py"""
-import pytest
+
 from gbif_registrar.utilities import read_registrations
 from gbif_registrar.register import get_local_dataset_group_id
 from gbif_registrar.register import get_local_dataset_endpoint
@@ -7,45 +7,6 @@ from gbif_registrar.register import get_gbif_dataset_uuid
 from gbif_registrar.register import register
 from gbif_registrar.register import request_gbif_dataset_uuid
 from gbif_registrar.config import PASTA_ENVIRONMENT
-
-
-@pytest.fixture(name="rgstrs")
-def rgstrs_fixture():
-    """Read the test registrations file into DataFrame fixture."""
-    return read_registrations("tests/registrations.csv")
-
-
-@pytest.fixture(name="local_dataset_id")
-def local_dataset_id_fixture():
-    """Create a local_dataset_id fixture for tests."""
-    # This local_dataset_id corresponds to a data package archived on EDI and
-    # formatted according to the DwCA-Event core.
-    return "edi.929.2"
-
-
-@pytest.fixture
-def local_dataset_group_id():
-    """Create a local_dataset_group_id fixture for tests."""
-    # This local_dataset_group_id corresponds to a data package archived on
-    # EDI and formatted according to the DwCA-Event core. This value is
-    # derived from the local_dataset_id fixture and should be the same as the
-    # local_dataset_id with the version number removed.
-    return "edi.929"
-
-
-@pytest.fixture(name="local_dataset_endpoint")
-def local_dataset_endpoint_fixture():
-    """Create a local_dataset_endpoint fixture for tests."""
-    # This local_dataset_endpoint corresponds to a data package archived on
-    # EDI and formatted according to the DwCA-Event core.
-    return "https://pasta.lternet.edu/package/download/eml/edi/929/2"
-
-
-@pytest.fixture(name="gbif_dataset_uuid")
-def gbif_dataset_uuid_fixture():
-    """Create a gbif_dataset_uuid fixture for tests."""
-    # This gbif_dataset_uuid is an example UUID returned by the GBIF API.
-    return "4e70c80e-cf22-49a5-8bf7-280994500324"
 
 
 def test_get_local_dataset_group_id(local_dataset_id):
